% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dire_bonjour.R
\name{dire_bonjour}
\alias{dire_bonjour}
\title{Dire bonjour}
\usage{
dire_bonjour(prenom = "toi")
}
\arguments{
\item{prenom}{un prénom en chaine de charactères}
}
\value{
une chaine de caractère qui dit bonjour au prénom spécifié
}
\description{
Dire bonjour
}
\examples{
dire_bonjour("Vincent")
dire_bonjour()

}
